2ファイル目以降の送信時にプログレスバーを 0 に戻していないのを修正した。 http://sourceforge.jp/ticket/browse.php?group_id=1412&tid=19671
git-svn-id: https://svn.osdn.net/svnroot/ttssh2/trunk@3703 f5f01b69-1e22-0410-acbf-894ab4bd6246

+++ teraterm/common/dlglib.c
@@ -110,7 +110,7 @@ void SetDlgPercent(HWND HDlg, int id_Item, int id_Progress, LONG a, LONG b, int 
 	_snprintf_s(NumStr,sizeof(NumStr),_TRUNCATE,"%3.1f%%",Num); 
 	SetDlgItemText(HDlg, id_Item, NumStr); 
 
-	if (id_Progress != 0 && p != NULL && *p >= 0 && *p < (int)Num) {
+	if (id_Progress != 0 && p != NULL && *p >= 0 && (double)*p < Num) {
 		*p = (int)Num;
 		SendMessage(GetDlgItem(HDlg, id_Progress), PBM_SETPOS, (WPARAM)*p, 0);
 	}
